7/requirements.txt
langgraph
langchain-openai
langchain-community
langchain-google-genai
langchain-groq
python-dotenv
pypdf      <-- برای خواندن PDF
faiss-cpu  <-- دیتابیس وکتوری (نسخه CPU)